fix(JSAM-68):move the actions and reducers in login folder
