+++ JSA-Mawsitsit-Frontend/src/components/App/AppReducer.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux';
 
 // Internal Dependencies
 import AccountManagementReducer from '../AccountManagement/reducers';
-import loginReducer from '../../reducers/loginReducer';
+import loginReducer from '../Login/reducers/loginReducer';
 
 const AppReducer = combineReducers({
   AccountManagement: AccountManagementReducer,
