+++ JSA-Mawsitsit-Frontend/src/containers/loginFormContainer.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import LoginForm from '../components/Login/LoginForm';
 
 // Internal Dependencies
-import { fetchLoginAction, updateLoginInfoAction } from '../actions/loginAction';
+import { fetchLoginAction, updateLoginInfoAction } from '../components/Login/actions/loginAction';
 
 const mapStateToProps = (state) => {
   const { loginInput, loginResponse } = state.loginForm;
